[#125 / #127] Support adding torrent from metadata by code
Before only the magnetlink route would actually work.

+++ bittorrent-extension-ut-metadata/src/main/java/org/johnnei/javatorrent/module/MetadataInformation.java
@@ -2,13 +2,13 @@ package org.johnnei.javatorrent.module;
 
 public class MetadataInformation {
 
-	private int metadataSize;
+	private long metadataSize;
 
-	public int getMetadataSize() {
+	public long getMetadataSize() {
 		return metadataSize;
 	}
 
-	public void setMetadataSize(int metadataSize) {
+	public void setMetadataSize(long metadataSize) {
 		this.metadataSize = metadataSize;
 	}
 }
